Refactor configuration manager into services
Large refactor that splits the monolithic ConfigurationManager into smaller services and registries. Introduces ConfigurationLoader, SerializationService, DeserializationService, ConfigurationSynchronizer, VersionManager and AdapterRegistry; ConfigurationManager now delegates to these services and exposes new TypeToken-based serializer APIs (registerTypeSerializer/deserialize). Adds a rich SerializationException for better error context and a FieldMeta record, moves several classes into service/model/integration packages, and centralizes logging via ConfigurationLogger. Also updates backup handling (ConfigBackupManager) to use the new logger and removes many inline I/O/serialization helpers from ConfigurationManager in favor of the new components.

+++ configuration/src/test/java/gloomlib/test/TypeInferenceBenchmark.java
@@ -2,7 +2,7 @@ package gloomlib.test;
 
 import gloomlib.configuration.ConfigurationFile;
 import gloomlib.configuration.ConfigurationPart;
-import gloomlib.configuration.TypeInference;
+import gloomlib.configuration.util.TypeInference;
 import org.junit.jupiter.api.*;
 
 import java.lang.reflect.Field;
@@ -220,7 +220,7 @@ public class TypeInferenceBenchmark {
             }
         }
         long hotTime = System.nanoTime() - hotStart;
-        long avgHotTime = hotTime / (BENCHMARK_ITERATIONS * fields.length);
+        long avgHotTime = hotTime / ((long) BENCHMARK_ITERATIONS * fields.length);
 
         System.out.printf("✓ Scanned %d fields%n", fields.length);
         System.out.printf("  - Cold cache (total): %,d ns%n", coldTime);
@@ -262,7 +262,7 @@ public class TypeInferenceBenchmark {
         System.out.println(TypeInference.getCacheStats());
         System.out.printf("  - Estimated memory: %,d bytes (%.2f KB)%n", memUsed, memUsed / 1024.0);
         System.out.printf("  - Memory per cache entry: ~%,d bytes%n",
-                memUsed / (fields.length * 3 + 3));
+                memUsed / (fields.length * 3L + 3));
     }
 
     // 测试配置类（包含各种复杂类型）
